More cleanup. remove cmd proj.
