+++ Fda.sln
@@ -39,8 +39,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HEC.FDA.Model", "HEC.FDA.Mo
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HEC.FDA.Statistics", "HEC.FDA.Statistics\HEC.FDA.Statistics.csproj", "{3292CA9D-B9A5-46BF-BF90-E71E394E5C4E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HEC.FDA.ModelCmd", "HEC.FDA.ModelCmd\HEC.FDA.ModelCmd.csproj", "{56449B16-DB55-4FBD-A57C-7C54F21C439D}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HEC.FDA.ModelTest", "HEC.FDA.ModelTest\HEC.FDA.ModelTest.csproj", "{9C53ED00-E9D3-4B21-9007-7B1A840CE00B}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HEC.FDA.StatisticsTest", "HEC.FDA.StatisticsTests\HEC.FDA.StatisticsTest.csproj", "{74BD5928-DE5D-43FD-AABF-27303F4F450E}"
@@ -191,18 +189,6 @@ Global
 		{3292CA9D-B9A5-46BF-BF90-E71E394E5C4E}.Release|x64.Build.0 = Release|Any CPU
 		{3292CA9D-B9A5-46BF-BF90-E71E394E5C4E}.Release|x86.ActiveCfg = Release|Any CPU
 		{3292CA9D-B9A5-46BF-BF90-E71E394E5C4E}.Release|x86.Build.0 = Release|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Debug|x64.Build.0 = Debug|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Debug|x86.Build.0 = Debug|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Release|x64.ActiveCfg = Release|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Release|x64.Build.0 = Release|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Release|x86.ActiveCfg = Release|Any CPU
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D}.Release|x86.Build.0 = Release|Any CPU
 		{9C53ED00-E9D3-4B21-9007-7B1A840CE00B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9C53ED00-E9D3-4B21-9007-7B1A840CE00B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9C53ED00-E9D3-4B21-9007-7B1A840CE00B}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -314,7 +300,6 @@ Global
 		{85D14B24-06D6-42D2-849D-E26456E41010} = {08C7D9FF-8AA8-465A-B851-E11D26E1CF01}
 		{151D1885-F68E-453F-A611-8E4AB273F573} = {960BBA43-9C13-478B-83B0-B1FB002DB14A}
 		{3292CA9D-B9A5-46BF-BF90-E71E394E5C4E} = {960BBA43-9C13-478B-83B0-B1FB002DB14A}
-		{56449B16-DB55-4FBD-A57C-7C54F21C439D} = {9C306E3C-BACC-45F9-BDAF-4A3D900EF9E4}
 		{9C53ED00-E9D3-4B21-9007-7B1A840CE00B} = {9C306E3C-BACC-45F9-BDAF-4A3D900EF9E4}
 		{74BD5928-DE5D-43FD-AABF-27303F4F450E} = {9C306E3C-BACC-45F9-BDAF-4A3D900EF9E4}
 		{DC66FA03-B7B5-47FF-8A4A-55F6370D23EF} = {960BBA43-9C13-478B-83B0-B1FB002DB14A}
