+++ HEC.FDA.Importer/HEC.FDA.Importer.csproj
@@ -1,11 +1,10 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
+	
   <PropertyGroup>
     <OutputType>Library</OutputType>
-	  <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-	  <OutputPath>..\Build\FDA2.0</OutputPath>
-    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
 	  <WarningLevel>0</WarningLevel> <!--!This turns off all warnings!-->
   </PropertyGroup>
+	
   <ItemGroup>
     <ProjectReference Include="..\HEC.FDA.Statistics\HEC.FDA.Statistics.csproj" />
   </ItemGroup>
