--- HEC.FDA.ModelCmd/HEC.FDA.ModelCmd.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<OutputType>Exe</OutputType>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-		<WarningLevel>0</WarningLevel> <!--!This turns off all warnings!-->
-	</PropertyGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\HEC.FDA.Model\HEC.FDA.Model.csproj" />
-		<ProjectReference Include="..\HEC.MVVMFramework.Model\HEC.MVVMFramework.Model.csproj" />
-	</ItemGroup>
-
-</Project>
