+++ HEC.FDA.Statistics/HEC.FDA.Statistics.csproj
@@ -1,7 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 	
 	<PropertyGroup>
-		<PackageId>HEC.FDA.Statistics</PackageId>
 		<Description>A library for computing statistics within HEC-FDA</Description>
 	</PropertyGroup>
