+++ HEC.MVVMFramework.ViewModel/HEC.MVVMFramework.ViewModel.csproj
@@ -1,10 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
-	</PropertyGroup>
-
-	<PropertyGroup>
 		<Description>A library providing functionality for MVVM ViewModels</Description>
 	</PropertyGroup>
