+++ ScratchSpace/EntryPoints/Nugent.cs
@@ -1,0 +1,7 @@
+﻿namespace ScratchSpace.EntryPoints;
+internal class Nugent
+{
+    public static void EntryPoint()
+    {
+    }
+}
